Changed some goals to ensure they return a scalar before minimize is called (future versions of numpy have problem in an array with one element is returned) dextranShape cut_front is now almost deterministic (at least to nearest 0.1 second) and far more accurate and stable than before due to other fixes it revealed that the cutting times where not always stable (I think this was due to non-smoothness at very small scales) and this propogated into parameter estimation and mcmc and this fixes the issue along with better detecting when the cuts can't be found improved logging of failed results for mcmc to show search space and parameter space

+++ CADETMatch/score.py
@@ -271,7 +271,7 @@ def find_cuts(times, values, spline, spline_der):
     max_time = times[max_index]
 
     def goal(time):
-        return -spline_der(time)
+        return -float(spline_der(time))
 
     result = scipy.optimize.minimize(goal, max_time, method="powell")
 
@@ -282,7 +282,7 @@ def find_cuts(times, values, spline, spline_der):
     min_time = times[min_index]
 
     def goal(time):
-        return abs(spline(time) - 1e-3 * max_value)
+        return float(abs(spline(time) - 1e-3 * max_value))
 
     result = scipy.optimize.minimize(goal, min_time, method="powell")
